+++ SMG_NothingIsAll/AIFoundation/AIThinkingControl/ThinkingUtils.h
@@ -14,7 +14,7 @@
 /**
  *  MARK:--------------------更新能量值--------------------
  */
-+(NSInteger) updateEnergy:(NSInteger)oriEnergy delta:(NSInteger)delta;
++(CGFloat) updateEnergy:(CGFloat)oriEnergy delta:(CGFloat)delta;
 
 @end
