+++ SMG_NothingIsAll/AIFoundation/AIThinkingControl/ThinkingUtils.m
@@ -28,7 +28,7 @@
 /**
  *  MARK:--------------------更新能量值--------------------
  */
-+(NSInteger) updateEnergy:(NSInteger)oriEnergy delta:(NSInteger)delta{
++(CGFloat) updateEnergy:(CGFloat)oriEnergy delta:(CGFloat)delta{
     oriEnergy += delta;
     return MAX(cMinEnergy, MIN(cMaxEnergy, oriEnergy));
 }
